+++ tilebox-workflows/tests/test_data.py
@@ -4,9 +4,11 @@ from tests.tasks_data import (
     automations,
     clusters,
     computed_tasks,
+    execution_stats,
     idling_responses,
     jobs,
     progress_indicators,
+    query_filters,
     storage_locations,
     task_identifiers,
     task_leases,
@@ -17,9 +19,11 @@ from tilebox.workflows.data import (
     AutomationPrototype,
     Cluster,
     ComputedTask,
+    ExecutionStats,
     Idling,
     Job,
     ProgressIndicator,
+    QueryFilters,
     StorageLocation,
     Task,
     TaskIdentifier,
@@ -46,6 +50,11 @@ def test_tasks_to_message_and_back(task: Task) -> None:
 @given(idling_responses())
 def test_idling_responses_to_message_and_back(idling: Idling) -> None:
     assert Idling.from_message(idling.to_message()) == idling
+
+
+@given(execution_stats())
+def test_execution_stats_to_message_and_back(execution_stats: ExecutionStats) -> None:
+    assert ExecutionStats.from_message(execution_stats.to_message()) == execution_stats
 
 
 @given(jobs())
@@ -87,3 +96,8 @@ def test_buckets_to_message_and_back(storage_location: StorageLocation) -> None:
 @given(automations())
 def test_automation_to_message_and_back(automation: AutomationPrototype) -> None:
     assert AutomationPrototype.from_message(automation.to_message()) == automation
+
+
+@given(query_filters())
+def test_query_filters_to_message_and_back(filters: QueryFilters) -> None:
+    assert QueryFilters.from_message(filters.to_message()) == filters