+++ tilebox-workflows/tilebox/workflows/jobs/client.py
@@ -9,6 +9,7 @@ from tilebox.datasets.query.time_interval import TimeInterval, TimeIntervalLike
 from tilebox.workflows.clusters.client import ClusterSlugLike, to_cluster_slug
 from tilebox.workflows.data import (
     Job,
+    JobState,
     QueryFilters,
     QueryJobsResponse,
 )
@@ -154,7 +155,13 @@ class JobClient:
         """
         return self._service.visualize(_to_uuid(job), direction, layout, sketchy)
 
-    def query(self, temporal_extent: TimeIntervalLike | IDIntervalLike, automation_id: UUID | None = None) -> list[Job]:
+    def query(
+        self,
+        temporal_extent: TimeIntervalLike | IDIntervalLike,
+        automation_ids: UUID | list[UUID] | None = None,
+        job_states: JobState | list[JobState] | None = None,
+        name: str | None = None,
+    ) -> list[Job]:
         """List jobs in the given temporal extent.
 
         Args:
@@ -170,11 +177,14 @@ class JobClient:
                 - tuple of two UUIDs: [start, end) -> Construct an IDInterval with the given start and end id
                 - tuple of two strings: [start, end) -> Construct an IDInterval with the given start and end id
                     parsed from the strings
-            automation_id: The automation id to filter jobs by. If specified, only jobs created by the given automation
-                are returned.
-
+            automation_ids: An automation id or list of automation ids to filter jobs by.
+                If specified, only jobs created by one of the selected automations are returned.
+            job_states: A job state or list of job states to filter jobs by. If specified, only jobs in one of the
+                selected states are returned.
+            name: A name to filter jobs by. If specified, only jobs with a matching name are returned. The match is
+                case-insensitive and uses a fuzzy matching scheme.
         Returns:
-            A list of jobs.
+            A list of jobs matching the given filters.
         """
         time_interval: TimeInterval | None = None
         id_interval: IDInterval | None = None
@@ -202,7 +212,21 @@ class JobClient:
                     end_inclusive=dataset_time_interval.end_inclusive,
                 )
 
-        filters = QueryFilters(time_interval=time_interval, id_interval=id_interval, automation_id=automation_id)
+        automation_ids = automation_ids or []
+        if not isinstance(automation_ids, list):
+            automation_ids = [automation_ids]
+
+        job_states = job_states or []
+        if not isinstance(job_states, list):
+            job_states = [job_states]
+
+        filters = QueryFilters(
+            time_interval=time_interval,
+            id_interval=id_interval,
+            automation_ids=automation_ids,
+            job_states=job_states,
+            name=name,
+        )
 
         def request(page: PaginationProtocol) -> QueryJobsResponse:
             query_page = Pagination(page.limit, page.starting_after)